Combine DisplayTest and TestRun (poor breaking up of logic)
