+++ client/components/TestRun/index.jsx
@@ -2,14 +2,28 @@ import React, { Component, Fragment } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { Helmet } from 'react-helmet';
-import { Col, Container, Row } from 'react-bootstrap';
+import { withRouter } from 'react-router-dom';
+import nextId from 'react-id-generator';
+import {
+    Button,
+    ButtonToolbar,
+    Col,
+    Container,
+    Modal,
+    Pagination,
+    Row
+} from 'react-bootstrap';
 import queryString from 'query-string';
 import {
     getActiveRunConfiguration,
     getActiveRuns,
     saveResult
 } from '../../actions/runs';
-import DisplayTest from '@components/DisplayTest';
+import RaiseIssueModal from '@components/RaiseIssueModal';
+import ReviewConflictsModal from '@components/ReviewConflictsModal';
+import StatusBar from '@components/StatusBar';
+import TestResult from '@components/TestResult';
+import TestIframe from '@components/TestIframe';
 import checkForConflict from '../../utils/checkForConflict';
 import './TestRun.css';
 
@@ -20,8 +34,12 @@ class TestRun extends Component {
         this.state = {
             currentTestIndex: 1,
             runComplete: false,
-            showTestNavigator: true
+            showTestNavigator: true,
+            showRaiseIssueModal: false,
+            showConfirmModal: false,
+            showConflictsModal: false
         };
+        this.buttonAction = null;
 
         this.displayNextTest = this.displayNextTest.bind(this);
         this.displayPreviousTest = this.displayPreviousTest.bind(this);
@@ -29,6 +47,28 @@ class TestRun extends Component {
         this.saveResultFromTest = this.saveResultFromTest.bind(this);
         this.deleteResultFromTest = this.deleteResultFromTest.bind(this);
         this.toggleTestNavigator = this.toggleTestNavigator.bind(this);
+
+        // Test run actions
+        this.handleSaveClick = this.handleSaveClick.bind(this);
+        this.handleNextTestClick = this.handleNextTestClick.bind(this);
+        this.handlePreviousTestClick = this.handlePreviousTestClick.bind(this);
+        this.handleCloseRunClick = this.handleCloseRunClick.bind(this);
+        this.handleRedoClick = this.handleRedoClick.bind(this);
+        this.handleEditClick = this.handleEditClick.bind(this);
+        this.handleRaiseIssueClick = this.handleRaiseIssueClick.bind(this);
+        this.handleConflictsModalClick = this.handleConflictsModalClick.bind(
+            this
+        );
+        this.handleRaiseIssueFromConflictClick = this.handleRaiseIssueFromConflictClick.bind(
+            this
+        );
+
+        // Modal actions
+        this.handleModalCloseClick = this.handleModalCloseClick.bind(this);
+        this.handleModalConfirmClick = this.handleModalConfirmClick.bind(this);
+
+        this.iframe = React.createRef();
+
     }
 
     async componentDidMount() {
@@ -102,6 +142,362 @@ class TestRun extends Component {
         return true;
     }
 
+    handleModalCloseClick() {
+        this.setState({
+            showConfirmModal: false
+        });
+    }
+
+    handleModalConfirmClick() {
+        const action = this.buttonAction;
+        this.buttonAction = null;
+        this.setState({
+            showConfirmModal: false
+        });
+        this.performButtonAction(action);
+    }
+
+    async performButtonAction(action) {
+        const {
+            openAsUser,
+            userId,
+            history,
+            run
+        } = this.props;
+
+        const testerId = openAsUser || userId;
+        const test = run.tests[this.state.currentTestIndex - 1];
+
+        switch (action) {
+            case 'saveTest': {
+                this.iframe.current.triggerSubmit();
+                break;
+            }
+            case 'closeTest': {
+                // Save the serialized form of iframe
+                if (this.iframe.current) {
+                    await this.iframe.current.processResults(null);
+                }
+                history.push(`/test-queue`);
+                break;
+            }
+            case 'goToNextTest': {
+                // Save the serialized form of iframe
+                if (this.iframe.current) {
+                    await this.iframe.current.processResults(null);
+                }
+                this.displayNextTest();
+                break;
+            }
+            case 'goToPreviousTest': {
+                // Save the serialized form of iframe
+                if (this.iframe.current) {
+                    await this.iframe.current.processResults(null);
+                }
+                this.displayPreviousTest();
+                break;
+            }
+            case 'redoTest': {
+                if (this.iframe.current) {
+                    this.iframe.current.reloadAndClear();
+                }
+                await this.deleteResultFromTest();
+                break;
+            }
+            case 'editTest': {
+                // save serialized form state, since it will be
+                // gone from state after results are deleted
+                const serializedForm = test.results[openAsUser].serialized_form;
+                await this.deleteResultFromTest();
+                // hydrate form with serialized state
+                this.iframe.current.reloadAndHydrate(serializedForm);
+                break;
+            }
+        }
+    }
+
+    confirm(action) {
+        this.buttonAction = action;
+        this.setState({
+            showConfirmModal: true
+        });
+    }
+
+    handleSaveClick() {
+        this.performButtonAction('saveTest');
+    }
+
+    handleNextTestClick() {
+        this.performButtonAction('goToNextTest');
+    }
+
+    handlePreviousTestClick() {
+        this.performButtonAction('goToPreviousTest');
+    }
+
+    handleRedoClick() {
+        this.confirm('redoTest');
+    }
+
+    handleCloseRunClick() {
+        this.performButtonAction('closeTest');
+    }
+
+    handleEditClick() {
+        this.performButtonAction('editTest');
+    }
+
+    handleRaiseIssueClick() {
+        this.setState({
+            showRaiseIssueModal: !this.state.showRaiseIssueModal,
+            showConflictsModal: false
+        });
+    }
+
+    handleConflictsModalClick() {
+        this.setState({
+            showConflictsModal: !this.state.showConflictsModal
+        });
+    }
+
+    handleRaiseIssueFromConflictClick() {
+        this.setState({
+            showConflictsModal: false,
+            showRaiseIssueModal: true
+        });
+    }
+
+    renderModal({ at_key, git_hash, run, test, testIndex, userId, testerId }) {
+
+        let modalTitle, action;
+
+        if (this.buttonAction === 'redoTest') {
+            modalTitle = 'Start Over';
+            action = `Are you sure you want to start over test ${testIndex}. Your progress (if any) will be lost.`;
+        }
+
+        return (
+            <>
+                <Modal
+                    show={this.state.showConfirmModal}
+                    onHide={this.handleModalCloseClick}
+                    centered
+                    animation={false}
+                >
+                    <Modal.Header closeButton>
+                        <Modal.Title>{modalTitle}</Modal.Title>
+                    </Modal.Header>
+                    <Modal.Body>{action}</Modal.Body>
+                    <Modal.Footer>
+                        <Button
+                            variant="secondary"
+                            onClick={this.handleModalCloseClick}
+                        >
+                            Cancel
+                        </Button>
+                        <Button
+                            variant="primary"
+                            onClick={this.handleModalConfirmClick}
+                        >
+                            Continue
+                        </Button>
+                    </Modal.Footer>
+                </Modal>
+
+                <ReviewConflictsModal
+                    onHide={this.handleConflictsModalClick}
+                    show={this.state.showConflictsModal}
+                    userId={userId}
+                    testerId={testerId}
+                    test={test}
+                    handleRaiseIssueClick={this.handleRaiseIssueClick}
+                />
+
+                <RaiseIssueModal
+                    at_key={at_key}
+                    git_hash={git_hash}
+                    onHide={this.handleRaiseIssueClick}
+                    run={run}
+                    show={this.state.showRaiseIssueModal}
+                    test={test}
+                    testIndex={testIndex}
+                    userId={userId}
+                    testerId={testerId}
+                />
+            </>
+        );
+    }
+
+    renderTest({ run, test, git_hash, at_key, testIndex, testerId }) {
+        const { conflicts } = this.state;
+        const { userId } = this.props;
+        this.testHasResult = test.results && test.results[testerId];
+
+        this.testResultsCompleted =
+            this.testHasResult && test.results[testerId].status === 'complete';
+
+        this.testResultsSkipped =
+            this.testHasResult &&
+            test.results[testerId].status === 'incomplete';
+
+        const statusProps = {
+            conflicts,
+            git_hash,
+            handleCloseRunClick: this.handleCloseRunClick,
+            handleNextTestClick: this.handleNextTestClick,
+            handlePreviousTestClick: this.handlePreviousTestClick,
+            handleRaiseIssueClick: this.handleRaiseIssueClick,
+            handleRedoClick: this.handleRedoClick,
+            handleConflictsModalClick: this.handleConflictsModalClick,
+            run,
+            test,
+            testIndex,
+            testerId
+        };
+
+        let testContent = null;
+
+        let primaryButtonGroup;
+
+        if (this.testResultsCompleted) {
+            primaryButtonGroup = (
+                <div className="testrun__button-toolbar-group">
+                    <Button variant="primary" onClick={this.handleNextTestClick}>
+                        Next test
+                    </Button>
+                    <Button variant="secondary" onClick={this.handleEditClick}>
+                        Edit results
+                    </Button>
+                </div>
+            );
+        } else {
+            primaryButtonGroup = (
+                <div className="testrun__button-toolbar-group">
+                    <Button variant="primary" onClick={this.handleSaveClick}>
+                        Save results
+                    </Button>
+                </div>
+            );
+        }
+
+        // note ButtonToolbar children are in row-reverse flex
+        // direction to align right when there is only one child
+        // and create a more logical tab order
+        let menuUnderContent = (
+            <ButtonToolbar className="testrun__button-toolbar">
+                {primaryButtonGroup}
+                <Pagination>
+                    <Pagination.First
+                        onClick={() => {
+                            this.displayTestByIndex(1);
+                        }}
+                    />
+                    <Pagination.Prev onClick={this.handlePreviousTestClick} />
+                    {run.tests.reduce((acc, t, i, arr) => {
+                        const item = (
+                            <Pagination.Item
+                                key={i}
+                                active={i + 1 === testIndex}
+                                onClick={() => {
+                                    this.displayTestByIndex(i + 1);
+                                }}
+                            >
+                                {i + 1}
+                            </Pagination.Item>
+                        );
+                        if (arr.length < 10 || i < 5 || i >= arr.length - 5) {
+                            acc.push(item);
+                        } else if (arr.length > 10 && i === 6) {
+                            acc.push(<Pagination.Ellipsis key={i} />);
+                        }
+                        return acc;
+                    }, [])}
+                    <Pagination.Next onClick={this.handleNextTestClick} />
+                    <Pagination.Last
+                        onClick={() => {
+                            this.displayTestByIndex(run.tests.length + 1);
+                        }}
+                    />
+                </Pagination>
+            </ButtonToolbar>
+        );
+        let menuRightOContent = (
+            <nav>
+                <Button
+                    className="btn-block"
+                    variant="primary"
+                    onClick={this.handleRaiseIssueClick}
+                >
+                    Raise an issue
+                </Button>
+                <Button
+                    className="btn-block"
+                    variant="primary"
+                    onClick={this.handleRedoClick}
+                >
+                    Start over
+                </Button>
+                <Button
+                    className="btn-block"
+                    variant="primary"
+                    onClick={this.handleCloseRunClick}
+                >
+                    Close
+                </Button>
+            </nav>
+        );
+
+        if (this.testResultsCompleted) {
+            testContent = <TestResult testResult={test.results[testerId]} />;
+        } else if (this.testResultsSkipped) {
+            testContent = (
+                <TestIframe
+                    git_hash={git_hash}
+                    file={test.file}
+                    at_key={at_key}
+                    onResults={this.saveResultFromTest}
+                    ref={this.iframe}
+                    serializedForm={test.results[testerId].serialized_form}
+                ></TestIframe>
+            );
+        } else {
+            testContent = (
+                <TestIframe
+                    git_hash={git_hash}
+                    file={test.file}
+                    at_key={at_key}
+                    onResults={this.saveResultFromTest}
+                    ref={this.iframe}
+                ></TestIframe>
+            );
+        }
+
+        let modals = this.renderModal({
+            at_key,
+            git_hash,
+            run,
+            test,
+            testIndex,
+            userId,
+            testerId
+        });
+
+        return (
+            <Fragment>
+                <h4 data-test="test-run-h4">Testing task: {test.name}</h4>
+                <StatusBar key={nextId()} {...statusProps} />
+                <Row>
+                    <Col md={9} className="test-iframe-contaner">
+                        <Row>{testContent}</Row>
+                        <Row>{menuUnderContent}</Row>
+                    </Col>
+                    <Col md={3}>{menuRightOContent}</Col>
+                </Row>
+                {modals}
+            </Fragment>
+        );
+    }
+
     render() {
         const {
             run,
@@ -163,21 +559,16 @@ class TestRun extends Component {
 
             if (!this.state.runComplete) {
                 testContent = (
-                    <DisplayTest
-                        key={`${test.id}/${this.state.currentTestIndex}`}
-                        run={run}
-                        test={test}
-                        testIndex={this.state.currentTestIndex}
-                        git_hash={git_hash}
-                        at_key={at_key}
-                        displayNextTest={this.displayNextTest}
-                        displayPreviousTest={this.displayPreviousTest}
-                        displayTestByIndex={this.displayTestByIndex}
-                        saveResultFromTest={this.saveResultFromTest}
-                        deleteResultFromTest={this.deleteResultFromTest}
-                        userId={userId}
-                        testerId={openAsUser || userId}
-                    />
+                    this.renderTest({
+                        key: `${test.id}/${this.state.currentTestIndex}`,
+                        run,
+                        test,
+                        testIndex: this.state.currentTestIndex,
+                        git_hash,
+                        at_key,
+                        userId,
+                        testerId: openAsUser || userId
+                    })
                 );
             } else {
                 content = <div>Tests are complete.</div>;
@@ -302,10 +693,11 @@ TestRun.propTypes = {
     activeRunConfiguration: PropTypes.object,
     dispatch: PropTypes.func,
     userId: PropTypes.number,
-    openAsUser: PropTypes.number,
+    openAsUsoer: PropTypes.number,
     run: PropTypes.object,
     testSuiteVersionData: PropTypes.object,
-    usersById: PropTypes.object
+    usersById: PropTypes.object,
+    history: PropTypes.object,
 };
 
 const mapStateToProps = (state, ownProps) => {
